test: move TerminologyResourceCollectionRestRequests to test.commons
