+++ tests/com.b2international.snowowl.test.commons/META-INF/MANIFEST.MF
@@ -28,6 +28,7 @@ Import-Package: org.slf4j;version="1.7.25"
 Export-Package: com.b2international.snowowl.test.commons,
  com.b2international.snowowl.test.commons.auth,
  com.b2international.snowowl.test.commons.codesystem,
+ com.b2international.snowowl.test.commons.collections,
  com.b2international.snowowl.test.commons.inject,
  com.b2international.snowowl.test.commons.rest,
  com.b2international.snowowl.test.commons.snomed,
