+++ tests/com.b2international.snowowl.test.commons/src/com/b2international/snowowl/test/commons/collections/TerminologyResourceCollectionRestRequests.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.b2international.snowowl.core.rest.resource;
+package com.b2international.snowowl.test.commons.collections;
 
 import static com.b2international.snowowl.test.commons.rest.RestExtensions.assertCreated;
 import static com.b2international.snowowl.test.commons.rest.RestExtensions.givenAuthenticatedRequest;
